+++ container-images/mds-transaction-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@container-images/mds-transactions-api",
+  "name": "@container-images/mds-transaction-api",
   "version": "0.0.1",
   "private": true,
   "description": "Docker Image for MDS Transactions API",
@@ -14,10 +14,10 @@
   ],
   "scripts": {
     "bundle": "webpack --mode=production --env npm_package_name=${npm_package_name} --env npm_package_version=${npm_package_version}",
-    "image": "pnpm bundle && ../../bin/build-helper buildImage mds-transactions-api:${npm_package_version}"
+    "image": "pnpm bundle && ../../bin/build-helper buildImage mds-transaction-api:${npm_package_version}"
   },
   "dependencies": {
-    "@mds-core/mds-transactions-service": "0.1.0",
+    "@mds-core/mds-transaction-api": "0.1.0",
     "@mds-core/mds-webpack-config": "0.1.0"
   }
 }
