+++ container-images/mds-transaction-service/webpack.config.ts
@@ -1,0 +1,3 @@
+import webpack from '@mds-core/mds-webpack-config'
+
+export default webpack.Bundle('../../packages/mds-transaction-service/server').UsingDefaultConfig()
