--- container-images/mds-transactions-service/webpack.config.ts
@@ -1,3 +1,0 @@
-import webpack from '@mds-core/mds-webpack-config'
-
-export default webpack.Bundle('../../packages/mds-transactions-service/server').UsingDefaultConfig()
