+++ packages/mds-transaction-api/handlers/get-transaction.ts
@@ -14,23 +14,25 @@
  * limitations under the License.
  */
 
-import { TransactionServiceClient, TransactionDomainModel } from '@mds-core/mds-transactions-service'
+import { TransactionServiceClient, TransactionDomainModel } from '@mds-core/mds-transaction-service'
+import { ApiRequestParams } from '@mds-core/mds-api-server'
 import express from 'express'
-import { TransactionApiRequest, TransactionApiResponse } from '../@types'
+import { TransactionApiResponse, TransactionApiRequest } from '../@types'
 
-export type TransactionApiCreateTransactionRequest = TransactionApiRequest<TransactionDomainModel>
+export type TransactionApiGetTransactionRequest = TransactionApiRequest & ApiRequestParams<'id'>
 
-export type TransactionApiCreateTransactionResponse = TransactionApiResponse<{ transaction: TransactionDomainModel }>
+export type TransactionApiGetTransactionResponse = TransactionApiResponse<{ transaction: TransactionDomainModel }>
 
-export const CreateTransactionHandler = async (
-  req: TransactionApiCreateTransactionRequest,
-  res: TransactionApiCreateTransactionResponse,
+export const GetTransactionHandler = async (
+  req: TransactionApiGetTransactionRequest,
+  res: TransactionApiGetTransactionResponse,
   next: express.NextFunction
 ) => {
   try {
-    const transaction = await TransactionServiceClient.createTransaction(req.body)
+    const { id } = req.params
+    const transaction = await TransactionServiceClient.getTransaction(id)
     const { version } = res.locals
-    return res.status(201).send({ version, transaction })
+    return res.status(200).send({ version, transaction })
   } catch (error) {
     next(error)
   }
