+++ packages/mds-transaction-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mds-core/mds-transactions-api",
+  "name": "@mds-core/mds-transaction-api",
   "version": "0.1.0",
   "description": "MDS Transaction API",
   "license": "Apache-2.0",
@@ -21,7 +21,7 @@
     "@mds-core/mds-api-server": "0.1.26",
     "@mds-core/mds-logger": "0.1.24",
     "@mds-core/mds-service-helpers": "0.1.0",
-    "@mds-core/mds-transactions-service": "0.1.0",
+    "@mds-core/mds-transaction-service": "0.1.0",
     "@mds-core/mds-types": "0.1.23",
     "@mds-core/mds-utils": "0.1.26",
     "@types/express": "4.17.10",
