--- packages/mds-transactions-api/handlers/get-operations.ts
@@ -1,42 +1,0 @@
-/**
- * Copyright 2020 City of Los Angeles
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { TransactionServiceClient } from '@mds-core/mds-transactions-service'
-import { TransactionOperationDomainModel } from '@mds-core/mds-transactions-service/@types'
-import { ApiRequestParams } from '@mds-core/mds-api-server'
-import express from 'express'
-import { TransactionApiRequest, TransactionApiResponse } from '../@types'
-
-export type TransactionApiGetTransactionOperationsRequest = TransactionApiRequest & ApiRequestParams<'transaction_id'>
-
-export type TransactionApiGetTransactionOperationsResponse = TransactionApiResponse<{
-  operations: TransactionOperationDomainModel[]
-}>
-
-export const GetTransactionOperationsHandler = async (
-  req: TransactionApiGetTransactionOperationsRequest,
-  res: TransactionApiGetTransactionOperationsResponse,
-  next: express.NextFunction
-) => {
-  try {
-    const { transaction_id } = req.params
-    const operations = await TransactionServiceClient.getTransactionOperations(transaction_id)
-    const { version } = res.locals
-    return res.status(200).send({ version, operations })
-  } catch (error) {
-    next(error)
-  }
-}
